requirements.txt
autopep8
django
django-rest-framework
django-cors-headers
